# Conflicts:
#	backend/consumer-management-8002/src/main/java/com/fellaverse/backend/controller/FeignFunctionManageController.java
#	backend/consumer-management-8002/src/main/java/com/fellaverse/backend/service/FeignFunctionManageService.java
#	backend/pom.xml
#	frontend/src/pages/admin/adminApp.jsx

+++ backend/common-api/src/main/java/com/fellaverse/backend/dto/UserBalanceStatusDTO.java
@@ -2,7 +2,6 @@ package com.fellaverse.backend.dto;
 
 import com.fellaverse.backend.annotation.EnumString;
 import com.fellaverse.backend.enumerator.UserStatus;
-import com.fellaverse.backend.validator.ValidGroup;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
@@ -16,7 +15,7 @@ public class UserBalanceStatusDTO implements Serializable {
     @NotNull(message = "User ID cannot be null")
     private Long id;
 
-    private Long wallet;
+    private Float wallet;
     @EnumString(value = {"normal","locked","unknown"}, message="invalid status")
     @NotNull(message = "User status cannot be null")
     private UserStatus status;
